Consistent naming of API registration function
