+++ origin/origin_api.go
@@ -20,16 +20,14 @@ package origin
 
 import (
 	"context"
-	"fmt"
-	"net/http"
+	"os"
+	"path/filepath"
 	"sync"
 	"time"
 
-	"github.com/gin-gonic/gin"
-	"github.com/pelicanplatform/pelican/common"
+	"github.com/pelicanplatform/pelican/config"
 	"github.com/pelicanplatform/pelican/metrics"
-	"github.com/pelicanplatform/pelican/token"
-	"github.com/pelicanplatform/pelican/token_scopes"
+	"github.com/pelicanplatform/pelican/param"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
 	"golang.org/x/sync/errgroup"
@@ -42,6 +40,37 @@ var (
 	notifyResponseOnce sync.Once
 	notifyChannel      chan bool
 )
+
+// Configure XrootD directory for both self-based and director-based file transfer tests
+func ConfigureXrootdMonitoringDir() error {
+	pelicanMonitoringPath := filepath.Join(param.Origin_RunLocation.GetString(),
+		"export", "pelican", "monitoring")
+
+	uid, err := config.GetDaemonUID()
+	if err != nil {
+		return err
+	}
+	gid, err := config.GetDaemonGID()
+	if err != nil {
+		return err
+	}
+	username, err := config.GetDaemonUser()
+	if err != nil {
+		return err
+	}
+
+	err = config.MkdirAll(pelicanMonitoringPath, 0755, uid, gid)
+	if err != nil {
+		return errors.Wrapf(err, "Unable to create pelican file trasnfer monitoring directory %v",
+			pelicanMonitoringPath)
+	}
+	if err = os.Chown(pelicanMonitoringPath, uid, -1); err != nil {
+		return errors.Wrapf(err, "Unable to change ownership of pelican file trasnfer monitoring directory %v"+
+			" to desired daemon user %v", pelicanMonitoringPath, username)
+	}
+
+	return nil
+}
 
 // Notify the periodic ticker that we have received a new response and it
 // should reset
@@ -84,53 +113,4 @@ func LaunchPeriodicDirectorTimeout(ctx context.Context, egrp *errgroup.Group) {
 			}
 		}
 	})
-}
-
-// The director periodically uploads/downloads files to/from all online
-// origins for testing. It sends a request reporting the status of the test result to this endpoint,
-// and we will update origin internal health status metric by what director returns.
-func directorTestResponse(ctx *gin.Context) {
-	status, ok, err := token.Verify(ctx, token.AuthOption{
-		Sources: []token.TokenSource{token.Header},
-		Issuers: []token.TokenIssuer{token.FederationIssuer},
-		Scopes:  []token_scopes.TokenScope{token_scopes.Pelican_DirectorTestReport},
-	})
-	if !ok {
-		ctx.JSON(status, gin.H{"error": err.Error()})
-	}
-
-	dt := common.DirectorTestResult{}
-	if err := ctx.ShouldBind(&dt); err != nil {
-		log.Errorf("Invalid director test response: %v", err)
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Invalid director test response: " + err.Error()})
-		return
-	}
-	// We will let the timer go timeout if director didn't send a valid json request
-	notifyNewDirectorResponse(ctx)
-	if dt.Status == "ok" {
-		metrics.SetComponentHealthStatus(metrics.OriginCache_Director, metrics.StatusOK, fmt.Sprintf("Director timestamp: %v", dt.Timestamp))
-		ctx.JSON(http.StatusOK, gin.H{"msg": "Success"})
-	} else if dt.Status == "error" {
-		metrics.SetComponentHealthStatus(metrics.OriginCache_Director, metrics.StatusCritical, dt.Message)
-		ctx.JSON(http.StatusOK, gin.H{"msg": "Success"})
-	} else {
-		log.Errorf("Invalid director test response, status: %s", dt.Status)
-		ctx.JSON(http.StatusBadRequest, gin.H{"error": fmt.Sprintf("Invalid director test response status: %s", dt.Status)})
-	}
-}
-
-// Configure API endpoints for origin that are not tied to UI
-func ConfigureOriginAPI(router *gin.Engine, ctx context.Context, egrp *errgroup.Group) error {
-	if router == nil {
-		return errors.New("Origin configuration passed a nil pointer")
-	}
-
-	metrics.SetComponentHealthStatus(metrics.OriginCache_Director, metrics.StatusWarning, "Initializing origin, unknown status for director")
-	// start the timer for the director test report timeout
-	LaunchPeriodicDirectorTimeout(ctx, egrp)
-
-	group := router.Group("/api/v1.0/origin-api")
-	group.POST("/directorTest", directorTestResponse)
-
-	return nil
 }
